+++ CHANGELOG.md
@@ -5,4 +5,5 @@
 #### Features
 
 - Scan the QR code to log in and save cookies in files.
-- Auto log in and leave comments (with count number and random letters)
+- Auto log in and leave comments (with count number and random letters).
+- Quit the current account and continue other accounts if the stored cookies are expired or submitting failed.